+++ src/main/java/de/fxnm/action/Login.java
@@ -44,8 +44,9 @@ public class Login extends BaseAction {
         final Presentation presentation = event.getPresentation();
         final ProjectStateService projectState = ProjectStateService.getService(event.getProject());
 
-        presentation.setEnabled(projectState.isServerConnectionEstablished()
-                && projectState.isManualLoginLogoutConfig());
+        presentation.setEnabled(
+                projectState.isServerConnectionEstablished()
+                        && projectState.isManualLoginLogoutConfig());
 
         if (ProjectStateService.getService(event.getProject()).isLoginConnectionEstablished()) {
             presentation.setIcon(PluginIcons.LOGIN_GREEN);