+++ src/main/java/de/fxnm/callable/account/service/LogInCallable.java
@@ -4,32 +4,22 @@ import com.intellij.credentialStore.Credentials;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.openapi.util.Pair;
 
-import org.jetbrains.annotations.NotNull;
-
-import java.awt.GridBagLayout;
 import java.io.IOException;
 import java.util.Objects;
-
-import javax.swing.JComponent;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JPasswordField;
-import javax.swing.JTextField;
 
 import de.fxnm.callable.BaseCallable;
 import de.fxnm.config.settings.password_safe.PasswordManager;
 import de.fxnm.exceptions.InternetConnectionException;
 import de.fxnm.exceptions.PasswordSafeException;
 import de.fxnm.exceptions.UsernamePasswordException;
+import de.fxnm.ui.account.LoginDialog;
 import de.fxnm.util.CodeTesterBundle;
 import de.fxnm.web.components.token.LoginToken;
 import de.fxnm.web.grabber.access_token.LoginTokenGrabber;
 
 import static de.fxnm.config.settings.password_safe.PasswordManager.LOGIN_DATE;
-import static de.fxnm.util.ComponentStatics.addComponent;
 
 public class LogInCallable extends BaseCallable<Boolean> {
 
@@ -66,7 +56,9 @@ public class LogInCallable extends BaseCallable<Boolean> {
 
                 super.finishedCallable("Login successful");
                 this.loginAbleWithSaveCredentials = true;
-
+            } catch (final UsernamePasswordException e) {
+                super.failedCallable("Login abort");
+                this.loginAbleWithSaveCredentials = false;
             } catch (final Throwable e) {
                 super.failedCallable("Failed to login");
                 LOG.warn(e);
@@ -94,42 +86,12 @@ public class LogInCallable extends BaseCallable<Boolean> {
     }
 
     private Pair<String, String> createLoginForm() throws UsernamePasswordException {
-        final UsernamePasswordDialog dialogWrapper = new UsernamePasswordDialog(CodeTesterBundle.message("plugin.runner.LoginRunner.Title"));
+        final LoginDialog dialogWrapper =
+                new LoginDialog(CodeTesterBundle.message("plugin.runner.LoginRunner.Title"));
         if (dialogWrapper.showAndGet()) {
             return dialogWrapper.getUserInput();
         } else {
             throw new UsernamePasswordException("No Username and Password provided");
-        }
-    }
-
-    private static final class UsernamePasswordDialog extends DialogWrapper {
-
-        private final JTextField userNameField = new JTextField();
-        private final JPasswordField passwordField = new JPasswordField();
-        private final JLabel userNameLabel = new JLabel(CodeTesterBundle.message("plugin.runner.LoginRunner.UserName"));
-        private final JLabel passwordLabel = new JLabel(CodeTesterBundle.message("plugin.runner.LoginRunner.Password"));
-
-        private UsernamePasswordDialog(final String title) {
-            super(true);
-            this.init();
-            this.setTitle(title);
-        }
-
-
-        @Override
-        protected @NotNull JComponent createCenterPanel() {
-            final JPanel panel = new JPanel(new GridBagLayout());
-
-            addComponent(this.userNameLabel, 0, 0, 1, 1, 1, 1, panel);
-            addComponent(this.userNameField, 1, 0, 2, 1, 1, 1, panel);
-            addComponent(this.passwordLabel, 0, 1, 1, 1, 1, 1, panel);
-            addComponent(this.passwordField, 1, 1, 2, 1, 1, 1, panel);
-
-            return panel;
-        }
-
-        public Pair<String, String> getUserInput() {
-            return new Pair<>(this.userNameField.getText(), new String(this.passwordField.getPassword()));
         }
     }
 }