+++ src/main/java/de/fxnm/config/ConfigurationListener.java
@@ -1,5 +1,7 @@
 package de.fxnm.config;
 
+import com.intellij.openapi.project.Project;
+
 import de.fxnm.web.components.category.Category;
 import de.fxnm.web.components.submission.SubmissionResult;
 
@@ -9,7 +11,7 @@ public interface ConfigurationListener {
 
     void displayInfoMessage(Boolean autoRemove, String message);
 
-    void displayCheckResult(SubmissionResult submissionResult);
+    void displayCheckResult(SubmissionResult submissionResult, Project project);
 
     void removeCheckResult();